Add cooldown

+++ app.js
@@ -2,16 +2,19 @@ require("dotenv").config();
 const fs = require("fs");
 const knex = require("./knex");
 const config = require("./config.json");
+const _ = require("lodash");
+const log = require("./commands/helpers/log");
 
 const Discord = require("discord.js");
 const client = new Discord.Client();
 client.commands = new Discord.Collection();
 
 var allGuildConfigs = {};
+var cooldownUsers = [];
+const cooldownTime = 1500;
 
 /* Listeners */
 const guildMemberAdd = require("./listeners/guildMemberAdd");
-const log = require("./commands/helpers/log");
 
 /* Attach commands to client */
 const commandFiles = fs
@@ -34,7 +37,7 @@ client.on("ready", async () => {
 
   /* Notify */
   // client.admin.send("I was offline, but I'm back now!");
-  client.user.setActivity(".help");
+  client.user.setActivity("say .help");
 });
 
 /* Clean up database once it leaves a server. */
@@ -67,9 +70,17 @@ client.on("message", async (msg) => {
   /* Catch DMs */
   /* DMs are reserved for setup */
   if (msg.channel.type === "dm") {
+    console.log(cooldownUsers);
+
     console.log(`${msg.author.username}: ${msg.content}`); // For feedback on how people _would_ use the DMs if they existed.
     if (msg.content.toLowerCase().startsWith(".help")) {
-      client.commands.get("help").execute(msg, false, false, client);
+      if (!cooldownUsers.includes(msg.author.id)) {
+        addCooldown(msg.author.id);
+        client.commands.get("help").execute(msg, false, false, client);
+      } else {
+        addCooldown(msg.author.id);
+        msg.reply("You're doing that too fast.");
+      }
     }
     return;
   }
@@ -139,33 +150,57 @@ client.on("message", async (msg) => {
 
   /* Commands */
   if (msg.content.startsWith(msg.channel.config.prefix)) {
-    /* Prepare arguments, attach to message. */
-    msg.content = msg.content.replace(/ +(?= )/g, ""); // Remove duplicate spaces
-    msg.content = msg.content.substring(msg.channel.config.prefix.length); // Remove prefix
-    msg.args = msg.content.split(" "); // Split into an arg array
-    msg.args[0] = msg.args[0].toLowerCase();
+    if (
+      msg.content.startsWith(
+        `${msg.channel.config.prefix}${msg.channel.config.prefix}`
+      )
+    ) {
+      return;
+    }
+    if (!cooldownUsers.includes(msg.author.id)) {
+      addCooldown(msg.author.id);
+      /* Prepare arguments, attach to message. */
+      msg.content = msg.content.replace(/ +(?= )/g, ""); // Remove duplicate spaces
+      msg.content = msg.content.substring(msg.channel.config.prefix.length); // Remove prefix
+      msg.args = msg.content.split(" "); // Split into an arg array
+      msg.args[0] = msg.args[0].toLowerCase();
 
-    /* If command exists, do it. */
-    const command = client.commands.get(msg.args[0]);
-    if (command) {
-      let allowed = false;
-      /* If command is locked */
-      if (command.privileged || command.facultyOnly) {
-        if (isModerator) allowed = true;
-        if (isFaculty && command.facultyOnly) allowed = true;
+      /* If command exists, do it. */
+      const command = client.commands.get(msg.args[0]);
+      if (command) {
+        let allowed = false;
+        /* If command is locked */
+        if (command.privileged || command.facultyOnly) {
+          if (isModerator) allowed = true;
+          if (isFaculty && command.facultyOnly) allowed = true;
+        } else {
+          /* If not locked */
+          allowed = true;
+        }
+        if (allowed) {
+          command.execute(msg, isModerator, isFaculty, client);
+        }
       } else {
-        /* If not locked */
-        allowed = true;
-      }
-      if (allowed) {
-        command.execute(msg, isModerator, isFaculty, client);
+        let sentMessage = await msg.channel.send(
+          `Bad command! Do \`.help\` for commands, ${msg.author}`
+        );
+        setTimeout(() => sentMessage.delete(), 3000);
       }
     } else {
-      const sentMessage = await msg.channel.send(
-        `Bad command! Do \`.help\` for commands, ${msg.author}`
+      const newMessage = await msg.channel.send(
+        `You're doing commands too fast, ${msg.author}! I need to take a breather.`
       );
-      setTimeout(() => sentMessage.delete(), 3000);
+      setTimeout(() => newMessage.delete(), 3000);
     }
   }
 });
+
+function addCooldown(ID) {
+  cooldownUsers.push(ID);
+  setTimeout(
+    () => _.pullAt(cooldownUsers, _.indexOf(cooldownUsers, ID)),
+    cooldownTime
+  );
+}
+
 client.login(process.env.CLIENT_TOKEN);
